Combine one or more 'esc' objects into a data frame

This method takes one or more objects of class esc (which are returned by each effect size calculation function) and returns the combined result as a single data frame. This can then be used for further computation, e.g. with the rma-function of the metafor-package.

combine_esc(...)

Arguments

...
One or more objects of class esc

Value

A data frame with all relevant information from the effect size calculation.

See also

Examples

e1 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40,
              grp2no = 45, study = "Study 1")
e2 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40, grp2no = 45,
              es.type = "or", study = "Study 2")
e3 <- esc_t(p = 0.03, grp1n = 100, grp2n = 150, study = "Study 3")
e4 <- esc_mean_sd(grp1m = 7, grp1sd = 2, grp1n = 50, grp2m = 9, grp2sd = 3,
                  grp2n = 60, es.type = "logit", study = "Study 4")

combine_esc(e1, e2, e3, e4)

#> # A tibble: 4 x 9
#>   study       es weight sample.size    se    var   ci.lo  ci.hi measure
#>   <chr>    <dbl>  <dbl>       <dbl> <dbl>  <dbl>   <dbl>  <dbl> <chr>  
#> 1 Study 1 -0.393   9.94         165 0.317 0.101  -1.01    0.228 logit  
#> 2 Study 2  0.675   9.94         165 0.317 0.101   0.363   1.26  or     
#> 3 Study 3  0.282  59.4          250 0.130 0.0168  0.0275  0.536 d      
#> 4 Study 4 -1.40    7.72         110 0.360 0.130  -2.10   -0.693 logit