Convert effect size d into f

Compute effect size f from effect size d.

convert_d2f(d, se, v, totaln, info = NULL, study = NULL)

Arguments

d
The effect size d.

se
The standard error of d. One of se or v must be specified.

v
The variance of d. One of se or v must be specified.

totaln
A vector of total sample size(s).

info
String with information on the transformation. Used for the print-method. Usually, this argument can be ignored

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

References

Cohen J. 1988. Statistical Power Analysis for the Behavioral Sciences. 2nd ed. Hillsdale, NJ: Erlbaum

Examples

# d to f
convert_d2f(d = 0.2, se = .1, totaln = 50)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>     Effect Size:   0.1000
#>  Standard Error:   0.1000
#>        Variance:   0.0100
#>        Lower CI:  -0.0960
#>        Upper CI:   0.2960
#>          Weight: 100.0000