Convert effect size d into correlation

Compute effect size correlation from effect size d.

esc_d2r(d, se, v, grp1n, grp2n, info = NULL, study = NULL)

convert_d2r(d, se, v, grp1n, grp2n, info = NULL, study = NULL)

Arguments

d
The effect size d.

se
The standard error of d. One of se or v must be specified.

v
The variance of d. One of se or v must be specified.

grp1n
Treatment group sample size.

grp2n
Control group sample size.

info
String with information on the transformation. Used for the print-method. Usually, this argument can be ignored

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln. Furthermore, Fisher's z and confidence intervals are returned.

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Examples

convert_d2r(d = 0.7, se = 0.5, grp1n = 70, grp2n = 80)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: effect size d to effect size correlation
#>     Effect Size:   0.3297
#>  Standard Error:   0.2420
#>        Variance:   0.0586
#>        Lower CI:  -0.1311
#>        Upper CI:   0.6734
#>          Weight:  17.0714
#>      Fisher's z:   0.3425
#>       Lower CIz:  -0.1319
#>       Upper CIz:   0.8169