Generate effect size data frame from other data

This method computes any effect size from raw values from a data frame. Convenient method to compute multiple effect sizes at once, when the required information to calculate effects sizes are stored in a table (i.e. data frame).

effect_sizes(data, ..., fun, es.type = c("d", "g", "or", "logit", "r",
  "f", "eta", "cox.or", "cox.log"))

Arguments

data
A data frame with columns that contain the values that are passed to one of the esc-functions.

...
Named arguments. The name (left-hand side) is the name of one of esc functions' argument, the argument (right-hand side) is the name of the column in data that holds the data values. See 'Examples'.

fun
Name of one of the esc-functions, as string, where arguments in ... are passed to. May either be the full function name (like "esc_t" or "esc_2x2") or the funcion name without the suffix "esc_" (like "t" or "2x2").

es.type
Type of effect size that should be returned.

"d"

returns standardized mean difference effect size d

"f"

returns effect size Cohen's f

"g"

returns adjusted standardized mean difference effect size Hedges' g

"or"

returns effect size as odds ratio

"cox.or"

returns effect size as Cox-odds ratio (see convert_d2or for details)

"logit"

returns effect size as log odds

"cox.log"

returns effect size as Cox-log odds (see convert_d2logit for details)

"r"

returns correlation effect size r

"eta"

returns effect size eta squared

Value

A data frame with the effect sizes computed for all data from data.

Details

This function rowwise iterates data and calls the function named in fun for the values taken from each row of data. The column names in data that contain the necessary values to compute the effect sizes should be passed as unquoted value for the arguments. The argument names should match those arguments for the esc-function that should be called from within effect_sizes().

Example:
If you want to compute effect sizes from chi-squared values, you would call esc_chisq(). This function name is used for the fun-argument: fun = "esc_chisq". esc_chisq() requires one of chisq or p as arguments, and totaln. Now data must have columns with values for either chisq or p, and effect_sizes() automatically selects the first non-missing value from data (see 'Examples').

Examples

tmp <- data.frame(
  tvalue = c(3.3, 2.9, 2.3),
  n = c(250, 200, 210),
  studyname = c("Study 1", "Study 2", "Study 3")
)
effect_sizes(tmp, t = tvalue, totaln = n, study = studyname, fun = "esc_t")

#> # A tibble: 3 x 9
#>   study      es weight sample.size    se    var  ci.lo ci.hi measure
#>   <chr>   <dbl>  <dbl>       <dbl> <dbl>  <dbl>  <dbl> <dbl> <chr>  
#> 1 Study 1 0.417   61.2         250 0.128 0.0163 0.167  0.668 d      
#> 2 Study 2 0.410   49.0         200 0.143 0.0204 0.130  0.690 d      
#> 3 Study 3 0.317   51.8         210 0.139 0.0193 0.0452 0.590 d      

# missing effect size results are dropped,
# shorter function name, calls "esc_t()"
tmp <- data.frame(
  tvalue = c(3.3, 2.9, NA, 2.3),
  n = c(250, 200, 210, 210),
  studyname = c("Study 1", "Study 2", NA, "Study 4")
)
effect_sizes(tmp, t = tvalue, totaln = n, study = studyname, fun = "t")

#> Warning: Row(s) 3 in `data` had missing values and were removed.

#> # A tibble: 3 x 9
#>   study      es weight sample.size    se    var  ci.lo ci.hi measure
#>   <chr>   <dbl>  <dbl>       <dbl> <dbl>  <dbl>  <dbl> <dbl> <chr>  
#> 1 Study 1 0.417   61.2         250 0.128 0.0163 0.167  0.668 d      
#> 2 Study 2 0.410   49.0         200 0.143 0.0204 0.130  0.690 d      
#> 3 Study 4 0.317   51.8         210 0.139 0.0193 0.0452 0.590 d      

tmp <- data.frame(
  coefficient = c(0.4, 0.2, 0.6),
  se = c(.15, .1, .2),
  treat = c(50, 60, 50),
  cntrl = c(45, 70, 40),
  author = c("Smith 2000", "Smith 2010 2", "Smith 2012")
)
effect_sizes(tmp, beta = coefficient, sdy = se, grp1n = treat, grp2n = cntrl,
    study = author, fun = "esc_beta", es.type = "or")

#> # A tibble: 3 x 9
#>   study           es weight sample.size    se   var ci.lo ci.hi measure
#>   <chr>        <dbl>  <dbl>       <dbl> <dbl> <dbl> <dbl> <dbl> <chr>  
#> 1 Smith 2000    4.80   6.59          95 0.390 0.152  2.24 10.3  or     
#> 2 Smith 2010 2  2.09   9.62         130 0.322 0.104  1.11  3.93 or     
#> 3 Smith 2012   15.0    5.30          90 0.434 0.189  6.40 35.1  or     

# the "esc_chisq" function requires *either* the chisq-argument *or*
# the pval-argument. If at least one of these values is present,
# effect size can be calculated. You can specify both arguments,
# and the first non-missing required value from "data" is taken.
tmp <- data.frame(
  chisqquared = c(NA, NA, 3.3, NA, 2.9),
  pval = c(.003, .05, NA, .12, NA),
  n = c(250, 200, 210, 150, 180),
  studyname = c("Study 1", "Study 2", "Study 3", "Study 4", "Study 5")
)
effect_sizes(tmp, chisq = chisqquared, p = pval, totaln = n,
             study = studyname, fun = "esc_chisq")

#> # A tibble: 5 x 9
#>   study      es weight sample.size    se    var     ci.lo ci.hi measure
#>   <chr>   <dbl>  <dbl>       <dbl> <dbl>  <dbl>     <dbl> <dbl> <chr>  
#> 1 Study 1 0.382   60.3         250 0.129 0.0166  1.30e- 1 0.635 d      
#> 2 Study 2 0.280   49.0         200 0.143 0.0204  5.55e-17 0.560 d      
#> 3 Study 3 0.253   51.7         210 0.139 0.0194 -1.99e- 2 0.525 d      
#> 4 Study 4 0.256   36.9         150 0.165 0.0271 -6.67e- 2 0.579 d      
#> 5 Study 5 0.256   44.3         180 0.150 0.0226 -3.86e- 2 0.550 d      

# if all required information are missing, data will be removed
tmp <- data.frame(
  chisqquared = c(NA, NA, 3.3, NA, NA),
  pval = c(.003, .05, NA, .12, NA),
  n = c(250, 200, 210, 150, 180),
  studyname = c("Study 1", "Study 2", "Study 3", "Study 4", "Study 5")
)
effect_sizes(tmp, chisq = chisqquared, p = pval, totaln = n,
             study = studyname, fun = "chisq")

#> Warning: Row(s) 5 in `data` had missing values and were removed.

#> # A tibble: 4 x 9
#>   study      es weight sample.size    se    var     ci.lo ci.hi measure
#>   <chr>   <dbl>  <dbl>       <dbl> <dbl>  <dbl>     <dbl> <dbl> <chr>  
#> 1 Study 1 0.382   60.3         250 0.129 0.0166  1.30e- 1 0.635 d      
#> 2 Study 2 0.280   49.0         200 0.143 0.0204  5.55e-17 0.560 d      
#> 3 Study 3 0.253   51.7         210 0.139 0.0194 -1.99e- 2 0.525 d      
#> 4 Study 4 0.256   36.9         150 0.165 0.0271 -6.67e- 2 0.579 d