Compute effect size from One-way Anova

Compute effect size from One-way Anova with two independent groups.

esc_f(f, totaln, grp1n, grp2n, es.type = c("d", "g", "or", "logit", "r", "f",
  "eta", "cox.or", "cox.log"), study = NULL)

Arguments

f
The F-value of the F-test.

totaln
Total sample size. Either totaln, or grp1n and grp2n must be specified.

grp1n
Treatment group sample size.

grp2n
Control group sample size.

es.type
Type of effect size that should be returned.

"d"

returns standardized mean difference effect size d

"f"

returns effect size Cohen's f

"g"

returns adjusted standardized mean difference effect size Hedges' g

"or"

returns effect size as odds ratio

"cox.or"

returns effect size as Cox-odds ratio (see esc_d2or for details)

"logit"

returns effect size as log odds

"cox.log"

returns effect size as Cox-log odds (see esc_d2logit for details)

"r"

returns correlation effect size r

"eta"

returns effect size eta squared

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

Note

This function only applies to one-way Anova F-tests with two independent groups, either equal or unequal sample sizes.

If es.type = "r", Fisher's transformation for the effect size r and their confidence intervals are also returned.

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Examples

# unequal sample size
esc_f(f = 5.5, grp1n = 100, grp2n = 150)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: F-value (one-way-Anova) to effect size d
#>     Effect Size:   0.3028
#>  Standard Error:   0.1298
#>        Variance:   0.0168
#>        Lower CI:   0.0483
#>        Upper CI:   0.5572
#>          Weight:  59.3472

# equal sample size
esc_f(f = 5.5, totaln = 200)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: F-value (one-way-Anova) to effect size d
#>     Effect Size:   0.3317
#>  Standard Error:   0.1424
#>        Variance:   0.0203
#>        Lower CI:   0.0526
#>        Upper CI:   0.6107
#>          Weight:  49.3218