Compute effect size from Mean Gain Scores and Standard Deviations

Compute effect size from Mean Gain Scores and Standard Deviations for pre-post tests.

esc_mean_gain(pre1mean, pre1sd, post1mean, post1sd, grp1n, gain1mean,
  gain1sd, grp1r, pre2mean, pre2sd, post2mean, post2sd, grp2n, gain2mean,
  gain2sd, grp2r, r, es.type = c("d", "g", "or", "logit", "r", "f",
  "eta", "cox.or", "cox.log"), study = NULL)

Arguments

pre1mean
The mean of the first group at pre-test.

pre1sd
The standard deviation of the first group at pre-test.

post1mean
The mean of the first group at post-test.

post1sd
The standard deviation of the first group at post-test.

grp1n
The sample size of the first group.

gain1mean
The mean gain between pre and post of the first group.

gain1sd
The standard deviation gain between pre and post of the first group.

grp1r
The (estimated) correlation of pre-post scores for the first group.

pre2mean
The mean of the second group at pre-test.

pre2sd
The standard deviation of the second group at pre-test.

post2mean
The mean of the second group at post-test.

post2sd
The standard deviation of the second group at post-test.

grp2n
The sample size of the second group.

gain2mean
The mean gain between pre and post of the second group.

gain2sd
The standard deviation gain between pre and post of the second group.

grp2r
The (estimated) correlation of pre-post scores for the second group.

r
Correlation for within-subject designs (paired samples, repeated measures).

es.type
Type of effect size that should be returned.

"d"

returns standardized mean difference effect size d

"f"

returns effect size Cohen's f

"g"

returns adjusted standardized mean difference effect size Hedges' g

"or"

returns effect size as odds ratio

"cox.or"

returns effect size as Cox-odds ratio (see convert_d2or for details)

"logit"

returns effect size as log odds

"cox.log"

returns effect size as Cox-log odds (see convert_d2logit for details)

"r"

returns correlation effect size r

"eta"

returns effect size eta squared

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

Details

For this function, either the gain scores of mean and sd (gain1mean and gain1sd for the first group and gain2mean and gain2sd for the second group) must be specified, or the pre-post values (pre1mean, post1mean, pre1sd and post1sd and the counterpart arguments for the second group).

If the pre-post standard deviations are available, no correlation value grp1r resp. grp2r needs to be specified, because these can then be computed based on t-value computation. However, if grp1r is specified, this value will be used (and no t-test performed).

Note

If es.type = "r", Fisher's transformation for the effect size r and their confidence intervals are also returned.

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Examples

# effect size of mean gain scores, with available pre-post values
esc_mean_gain(pre1mean = 13.07, pre1sd = 11.95, post1mean = 6.1,
              post1sd = 8.33, grp1n = 78, pre2mean = 10.77, pre2sd = 10.73,
              post2mean = 8.83, post2sd = 9.67, grp2n = 83)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: mean gain score to effect size d
#>     Effect Size:   0.4905
#>  Standard Error:   0.3194
#>        Variance:   0.1020
#>        Lower CI:  -0.1356
#>        Upper CI:   1.1165
#>          Weight:   9.8001

# same as above, but with assumed correlation of .5
# Note that effect size is the same, but variance differs
esc_mean_gain(pre1mean = 13.07, pre1sd = 11.95, post1mean = 6.1, grp1r = .5,
              post1sd = 8.33, grp1n = 78, pre2mean = 10.77, pre2sd = 10.73,
              post2mean = 8.83, post2sd = 9.67, grp2n = 83, grp2r = .5)

#> Pre-post correlation `grp1r` is specified, although it could be computed from `pre1sd` and `post1sd`. If `grp1r` is missing, correlation of pre-post scores will be computed automatically.FALSE

#> Pre-post correlation `grp2r` is specified, although it could be computed from `pre2sd` and `post2sd`. If `grp2r` is missing, correlation of pre-post scores will be computed automatically.FALSE

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: mean gain score to effect size d
#>     Effect Size:   0.4905
#>  Standard Error:   0.1600
#>        Variance:   0.0256
#>        Lower CI:   0.1768
#>        Upper CI:   0.8041
#>          Weight:  39.0385

# effect size based on gain scores for mean and sd. note that the
# pre-post correlations must be given
esc_mean_gain(gain1mean = 1.5, gain1sd = 1, grp1n = 40, grp1r = .5,
              gain2mean = .7, gain2sd = .8, grp2n = 50, grp2r = .5)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: mean gain score to effect size d
#>     Effect Size:   0.8947
#>  Standard Error:   0.2224
#>        Variance:   0.0494
#>        Lower CI:   0.4589
#>        Upper CI:   1.3305
#>          Weight:  20.2237