Compute effect size from Mean and Standard Deviation

Compute effect size from mean and either group-based standard deviations or full sample standard deviation.

esc_mean_sd(
  grp1m,
  grp1sd,
  grp1n,
  grp2m,
  grp2sd,
  grp2n,
  totalsd,
  r,
  es.type = c("d", "g", "or", "logit", "r", "cox.or", "cox.log"),
  study = NULL
)

Arguments

grp1m
The mean of the first group.

grp1sd
The standard deviation of the first group.

grp1n
The sample size of the first group.

grp2m
The mean of the second group.

grp2sd
The standard deviation of the second group.

grp2n
The sample size of the second group.

totalsd
The full sample standard deviation. Either grp1sd and grp2sd, or totalsd must be specified.

r
Correlation for within-subject designs (paired samples, repeated measures).

es.type
Type of effect size that should be returned.

"d"

returns standardized mean difference effect size d

"f"

returns effect size Cohen's f

"g"

returns adjusted standardized mean difference effect size Hedges' g

"or"

returns effect size as odds ratio

"cox.or"

returns effect size as Cox-odds ratio (see convert_d2or for details)

"logit"

returns effect size as log odds

"cox.log"

returns effect size as Cox-log odds (see convert_d2logit for details)

"r"

returns correlation effect size r

"eta"

returns effect size eta squared

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

Note

If es.type = "r", Fisher's transformation for the effect size r and their confidence intervals are also returned.

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Examples

# with standard deviations for each group
esc_mean_sd(
  grp1m = 7, grp1sd = 2, grp1n = 50,
  grp2m = 9, grp2sd = 3, grp2n = 60,
  es.type = "logit"
)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: mean and sd to effect size logits
#>     Effect Size:  -1.3982
#>  Standard Error:   0.3599
#>        Variance:   0.1295
#>        Lower CI:  -2.1035
#>        Upper CI:  -0.6928
#>          Weight:   7.7211

# effect-size d, within-subjects design
esc_mean_sd(
  grp1m = 7, grp1sd = 2, grp1n = 50,
  grp2m = 9, grp2sd = 3, grp2n = 60, r = .7
)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: mean and sd (within-subject) to effect size d
#>     Effect Size:  -0.9325
#>  Standard Error:   0.2015
#>        Variance:   0.0406
#>        Lower CI:  -1.3275
#>        Upper CI:  -0.5375
#>          Weight:  24.6189

# with full sample standard deviations
esc_mean_sd(grp1m = 7, grp1n = 50, grp2m = 9, grp2n = 60, totalsd = 4)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: mean and sd to effect size d
#>     Effect Size:  -0.5164
#>  Standard Error:   0.1946
#>        Variance:   0.0379
#>        Lower CI:  -0.8979
#>        Upper CI:  -0.1350
#>          Weight:  26.4000