Convert effect size OR from d

Compute effect size d from effect size OR.

esc_or2d(or, se, v, totaln, es.type = c("d", "cox.d", "g", "f", "eta"),
  info = NULL, study = NULL)

Arguments

or
The effect size as odds ratio.

se
The standard error of d. One of se or v must be specified.

v
The variance of d. One of se or v must be specified.

totaln
A vector of total sample size(s).

es.type
Type of effect size that should be returned.

"d"

returns effect size d

"cox.d"

returns effect size d, based on Cox method

"g"

returns effect size Hedges' g (see hedges_g)

info
String with information on the transformation. Used for the print-method. Usually, this argument can be ignored

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

Note

While or is the exponentiated log odds, the variance or standard error need to be on the log-scale!

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Examples

esc_or2d(3.56, se = 0.91)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: effect size OR to effect size d
#>     Effect Size:   0.7001
#>  Standard Error:   0.5017
#>        Variance:   0.2517
#>        Lower CI:  -0.2833
#>        Upper CI:   1.6834
#>          Weight:   3.9728

esc_d2or(0.7, se = 0.5)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: effect size d to effect size OR
#>     Effect Size:   3.5596
#>  Standard Error:   0.9069
#>        Variance:   0.8225
#>        Lower CI:   0.6018
#>        Upper CI:  21.0553
#>          Weight:   1.2159