Compute effect size from Phi coefficient

Compute effect size from phi coefficient

esc_phi(
  phi,
  p,
  totaln,
  es.type = c("d", "g", "or", "logit", "r", "f", "eta", "cox.or", "cox.log"),
  study = NULL
)

Arguments

phi
The phi value. One of phi or p must be reported.

p
The p-value of the chi-squared or phi-value.

totaln
A vector of total sample size(s).

es.type
Type of effect size that should be returned.

"d"

returns standardized mean difference effect size d

"f"

returns effect size Cohen's f

"g"

returns adjusted standardized mean difference effect size Hedges' g

"or"

returns effect size as odds ratio

"cox.or"

returns effect size as Cox-odds ratio (see convert_d2or for details)

"logit"

returns effect size as log odds

"cox.log"

returns effect size as Cox-log odds (see convert_d2logit for details)

"r"

returns correlation effect size r

"eta"

returns effect size eta squared

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

Note

This effect size should only be used for data from 2x2 frequency tables. Furthermore, use this approximation for the effect size only, if information about the 2x2 frequencies or proportions are not available. Else, esc_2x2 or esc_bin_prop provide better estimates for the effect size.

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Examples

# Effect size based on chi-squared value
esc_phi(phi = .67, totaln = 100)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: phi-value to effect size d
#>     Effect Size:   1.8051
#>  Standard Error:   0.2694
#>        Variance:   0.0726
#>        Lower CI:   1.2770
#>        Upper CI:   2.3331
#>          Weight:  13.7775

# Effect size based on p-value of chi-squared
esc_phi(p = .003, totaln = 100)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: chi-squared-value to effect size d
#>     Effect Size:   0.6215
#>  Standard Error:   0.2094
#>        Variance:   0.0439
#>        Lower CI:   0.2111
#>        Upper CI:   1.0320
#>          Weight:  22.7981