Write one or more 'esc' objects into an Excel csv-file

This method is a small wrapper to write csv-files. It writes the results from combine_esc into an Excel csv-file.

write_esc(..., path, sep = ",")

Arguments

...
One or more objects of class esc

path
Path to write to, or just file name (to write to working directory).

sep
The field separator string. In some Western European locales, Excel uses a semicolon by default, while in other locales the field separator string in Excel is a comma.

Value

Invisibly returns the combined data frame that is written to the csv-file (see combine_esc).

Note

For Western European locales, the sep-argument probably needs to be set to semicolon (sep = ";"), so Excel reads the csv-file properly. If sep = ";", write.csv2 is used to write the file. Else, write_excel_csv is used.

See also

Examples

e1 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40,
              grp2no = 45, study = "Study 1")
e2 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40, grp2no = 45,
              es.type = "or", study = "Study 2")
e3 <- esc_t(p = 0.03, grp1n = 100, grp2n = 150, study = "Study 3")
e4 <- esc_mean_sd(grp1m = 7, grp1sd = 2, grp1n = 50, grp2m = 9, grp2sd = 3,
                  grp2n = 60, es.type = "logit", study = "Study 4")

# write to current working directory,
# file extension ".csv" is automatically added
# NOT RUN {
write_esc(e1, e2, e3, e4, path = "EffSizes")
# }