Combine one or more 'esc' objects into a data frame

This method takes one or more objects of class esc (which are returned by each effect size calculation function) and returns the combined result as a single data frame. This can then be used for further computation, e.g. with the rma-function of the metafor-package.

combine_esc(...)

Arguments

...
One or more objects of class esc

Value

A data frame with all relevant information from the effect size calculation.

See also

Examples

e1 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40,
              grp2no = 45, study = "Study 1")
e2 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40, grp2no = 45,
              es.type = "or", study = "Study 2")
e3 <- esc_t(p = 0.03, grp1n = 100, grp2n = 150, study = "Study 3")
e4 <- esc_mean_sd(grp1m = 7, grp1sd = 2, grp1n = 50, grp2m = 9, grp2sd = 3,
                  grp2n = 60, es.type = "logit", study = "Study 4")

combine_esc(e1, e2, e3, e4)

#>     study         es    weight sample.size        se        var       ci.lo
#> 1 Study 1 -0.3930426  9.944751         165 0.3171050 0.10055556 -1.01455689
#> 2 Study 2  0.6750000  9.944751         165 0.3171050 0.10055556  0.36256305
#> 3 Study 3  0.2817789 59.433720         250 0.1297130 0.01682547  0.02754605
#> 4 Study 4 -1.3981827  7.721145         110 0.3598812 0.12951447 -2.10353685
#>        ci.hi measure
#> 1  0.2284717   logit
#> 2  1.2566780      or
#> 3  0.5360117       d
#> 4 -0.6928285   logit