Convert effect size d into OR

Compute effect size OR from effect size d.

esc_d2or(d, se, v, totaln, es.type = c("logit", "cox"), info = NULL,
  study = NULL)

convert_d2or(d, se, v, totaln, es.type = c("logit", "cox"),
  info = NULL, study = NULL)

Arguments

d
The effect size d.

se
The standard error of d. One of se or v must be specified.

v
The variance of d. One of se or v must be specified.

totaln
A vector of total sample size(s).

es.type
Type of effect size odds ratio that should be returned. May be es.type = "logit" or es.type = "cox" (see 'Details').

info
String with information on the transformation. Used for the print-method. Usually, this argument can be ignored

study
Optional string with the study name. Using combine_esc or as.data.frame on esc-objects will add this as column in the returned data frame.

Value

The effect size es, the standard error se, the variance of the effect size var, the lower and upper confidence limits ci.lo and ci.hi, the weight factor w and the total sample size totaln.

Details

Conversion from d to odds ratios can be done with two methods:

es.type = "logit"

uses the Hasselblad and Hedges logit method.

es.type = "cox"

uses the modified logit method as proposed by Cox. This method performs slightly better for rare or frequent events, i.e. if the success rate is close to 0 or 1.

Note

Effect size is returned as exp(log_values) (odds ratio), confidence intervals are also exponentiated. To get the log-values, use convert_d2logit. However, variance and standard error of this function are returned on the log-scale!

References

Lipsey MW, Wilson DB. 2001. Practical meta-analysis. Thousand Oaks, Calif: Sage Publications

Wilson DB. 2016. Formulas Used by the "Practical Meta-Analysis Effect Size Calculator". Unpublished manuscript: George Mason University

Cox DR. 1970. Analysis of binary data. New York: Chapman & Hall/CRC

Hasselblad V, Hedges LV. 1995. Meta-analysis of screening and diagnostic tests. Psychological Bulletin 117(1): 167–178. doi: 10.1037/0033-2909.117.1.167

Borenstein M, Hedges LV, Higgins JPT, Rothstein HR. 2009. Introduction to Meta-Analysis. Chichester, West Sussex, UK: Wiley

Examples

# d to odds ratio
convert_d2or(0.7, se = 0.5)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: effect size d to effect size OR
#>     Effect Size:   3.5596
#>  Standard Error:   0.9069
#>        Variance:   0.8225
#>        Lower CI:   0.6018
#>        Upper CI:  21.0553
#>          Weight:   1.2159

# odds ratio to d
convert_or2d(3.56, se = 0.91)

#> 
#> Effect Size Calculation for Meta Analysis
#> 
#>      Conversion: effect size OR to effect size d
#>     Effect Size:   0.7001
#>  Standard Error:   0.5017
#>        Variance:   0.2517
#>        Lower CI:  -0.2833
#>        Upper CI:   1.6834
#>          Weight:   3.9728