Effect Size Computation for Meta Analysis

This is an R implementation of the web-based 'Practical Meta-Analysis Effect Size Calculator' from David B. Wilson.

Based on the input, the effect size can be returned as standardized mean difference (d), Hedges' g, correlation coefficient effect size r or Fisher's transformation z, odds ratio or log odds effect size.

Return values

The return value of all functions has the same structure:

The effect size, whether being d, g, r, (Cox) odds ratios or (Cox) logits, is always named es.

The standard error of the effect size, se.

The variance of the effect size, var.

The lower and upper confidence limits ci.lo and ci.hi.

The weight factor, based on the inverse-variance, w.

The total sample size totaln.

The effect size measure, measure, which is typically specified via the es.type-argument.

Information on the effect-size conversion, info.

A string with the study name, if the study-argument was specified in function calls.

Correlation Effect Size

If the correlation effect size r is computed, the transformed Fisher's z and their confidence intervals are also returned. The variance and standard error for the correlation effect size r are always based on Fisher's transformation.

Odds Ratio Effect Size

For odds ratios, the variance and standard error are always returned on the log-scale!

Preparing an Effect Size Data Frame for Meta-Analysis

The results of the effect size calculation functions in this package are returned as list with a esc-class attribute. The combine_esc-function takes one or more of these esc-objects and combines them into a data.frame that can be used as argument for further use, for instance with the rma-function.

    e1 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40,
                  grp2no = 45, study = "Study 1")
    e2 <- esc_2x2(grp1yes = 30, grp1no = 50, grp2yes = 40, grp2no = 45,
                  es.type = "or", study = "Study 2")
    e3 <- esc_t(p = 0.03, grp1n = 100, grp2n = 150, study = "Study 3")
    e4 <- esc_mean_sd(grp1m = 7, grp1sd = 2, grp1n = 50, grp2m = 9, grp2sd = 3,
                      grp2n = 60, es.type = "logit", study = "Study 4")

    mydat <- combine_esc(e1, e2, e3, e4)

    metafor::rma(yi = es, sei = se, method = "REML", data = mydat)